requirements.txt
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
colorzero==2.0
distro==1.8.0
gpiozero==2.0.1
idna==3.3
lgpio==0.2.2.0
paho-mqtt==1.6.1
pigpio==1.78
pycryptodomex==3.11.0
python-apt==2.6.0
python-dotenv==0.21.0
requests==2.28.1
rpi-lgpio==0.6
six==1.16.0
smbus2==0.4.2
spidev==3.5
ssh-import-id==5.10
toml==0.10.2
urllib3==1.26.12
